requirements.txt
fastapi
uvicorn
python-dotenv
supabase
psycopg2-binary
pydantic
SQLAlchemy
requests
langchain
langchain-community
langchain-ollama
chromadb
openai
tiktoken
sentence-transformers
pyjwt
passlib==1.7.4
bcrypt==4.0.1
wandb


